Code Complete
I had to reset all the files to their initial state, meaning without any modifications, then try again to do the change to redux.

Thanks to the video from last class that was not available when I first tried to do the homework, I was able to follow the step by step of how to change from context to redux, and do the required modification to the files to be able to make the app work with redux.